Move Batch4's passing UserDefinedStructTests into Batch1
