+++ Tests/Batch1/Constants.cs
@@ -37,6 +37,7 @@
         public const string MODULE_REGEX_JS = "RegexJS";
         public const string MODULE_ENUM = "Enum";
         public const string MODULE_GUID = "Guid";
+        public const string MODULE_STRUCT = "Struct";
 
         public const string MODULE_MATH = "Math";
         public const string MODULE_DECIMAL_MATH = "Decimal Math";
